\name{Bahrami} 
\docType{data} 
\alias{Bahrami}
\title{Body Mass Index and Steatosis Data} 
\description{The medical data set is obtained from an observational study in
 the Taleghani  hospital in Tehran. These data record the 
Steatosis and BMI for 15 diabetic patients. Steatosis is the
process describing the abnormal retention of lipids within a cell.
It reflects an impair of the normal process of synthesis and
breakdown of triglyceride fat. Excess lipid accumulates in vescles
that displace the cytoplasm. BMI is a statistical measure of the
weight of body mass index. A person scaled height body mass index
may be accurately calculated using any of the formulas such as
 BMI=w/H^2 where W is weight and H is height. As
explanatory variable for both Steatosis and BMI, the systolic
blood pressure (SBP) is defined as the peak pressure in the
arteries, which occurs near the beginning of the cardiac cycle.
The normal rate, in adult humans, for systolic is near but less
than 120 mmHg. As another explanatory variable, duration of diabet
is an amount of time or a particular time interval which a person
take diabet (a metabolic
 disorder characterized by high blood suger and other signs). Two
 more explanatory variables are age and gender.} 
\usage{data(Bahrami)} 
\format{A data frame with 15 observations on the following 6 variables:AGE,DURATION,SBP,STEATOS,BMI,GENDER} 
\source{The medical data set is obtained from an observational study in
 the Taleghani  hospital in Tehran.} 
\references{Bahrami Samani, E.,  Ganjali, M. and Khodaddadi, A.  (2008). A Latent Variable Model for Mixed Continuous and Ordinal Responses. Journal of Statistical Theory and Applications. 7(3):337-349. } 
\keyword{datasets}